Partial bug Fix # 116214 Inconvenient keyboard navigation in members hierarchy
Fixed the issue 1. "when pressing the down arrow key and holding it, nothing happens until you release it, and then the selection only moves one item downwards. I expected it to work as in any other tree, moving many members, when you press the key and hold it."

This was fixed by registering the action on key press rather than key release.