Improve code completion: include the currently called method surrounding the caret if you're in the parameter list.  Also use this element as the documentation element for parameter completion. Improve the parameter completion I committed yesterday a bit (e.g. @config should only apply to the most recently mentioned @param tag, and fix some issues with computing the right parameter).   Improve the Go To Declaration code which picks which of many similar functions should be chosen (it wasn't working right with functions both inherited and overridden).
Add unit tests for all this. I've finally figured out how to force the
index to be updated for the current file as well, which makes code completion
for local functions and classes finally working in unit tests. Update both
JavaScript and Ruby unit tests to add regression tests for these scenarios.