Issue 147459 - edge labels can't be moved with ctrl-arrow Issue 146861 - moving link label with shortcut keys shifts diagram instead of moving the label
A dotted line will now appear between the connection widget and the able.  The dotted line attaches to the labels anchor point.  For example: on a multiplicity label the dotted line will connect to either the source or target location depending to if the label is associated with the source or target end.