processes termination-related tests fix ...
It would be better to separate termination with TERM and with KILL and
have different methods for this (as Ivan suggested).
This likely to be done in a new API...

So.. but for now KILL is send to a process if TERM had no effect for some time.
(This is how it worked before). But this could be not what is expected by user...