Wrap jsch channel's output stream to make a window between "is-ok-to-write" check and write operation as small as possible.
Unfortunately it is not possible to eliminate this check-then-ack pattern in this case and
there is an unavoidable race condition here. And the bigger problem that jsch closes the whole Session
if an attempt to write to not connected channel is performed. So this is an attempt to make the
window of vulnerability as small as possible.