209028 - Trailing spaces in a file are not recognized as a part of namespace
Improved processing of trailing spaces in a PHP file without ending delimiter, spaces are added to a Program node. Can't be added to a NamespaceDeclaration node properly, because when ND node is created, noone knows if there will be just another ND after it, so noone knows whether the trailing spaces should be added to this or that. But all features seems to be working right now.