#244286: Moving a project does not preserve file and folder timestamps
The copy operation is used internally during move. To recognize that the copy
is called from a move (without API changes), new thread local variable was
introduced. If this variable is set to true, the last-modified date of the target
file is set after copying.

Using NIO2, which could make the implementation simpler, should be considered
for the next release.