Improve performance during svn cache init.
#250295 kind of revealed a performance bottleneck during cache init. With a lot of modified files
the IDE is stuck simply only because it logs the modifications so we know about it when reading the IDE log.

Instead of logging the files directly the log contains just the note about the total number of files
and a hint that the exact files/folders may be enabled using a system property.