#253645 - org.netbeans.modules.maven.NbArtifactFixer.resolve: LowPerformance took 20204 ms. Do not attempt to create VCSContext while projects are being loaded.
Building VCSContext may be blocked by opening projects task, which in AWT may freeze IDE
for tens of seconds. I am trying an experimental fix that leaves git actions enabled for the
time projects are loading and only when invoked the actions check their enable status in the
complete range.