#256890: When selecting 'View Data' the resulting SQL statement has no limit
Many JDCB drivers don't place sensible limits on queries, if the limits are not part of
the SQL itself. This changeset generates the select with TOP/FIRST/LIMIT/FETCH FIRST
modification. The implementation tries to cover most dbs, as a fallback a plain
SQL is generated if DB is not recognised.