#262220: java.sql.SQLException: no database metadata available
Log exception when properties of a DB connection can't be updated on the
INFO level to the message log. This replaces logging this on level SEVERE
and generating mostly useless bug reports (exception reporter).

If a real bug happens afterwards, the hope is, that a real bug report is
filed.